+++ src/sites/reklama.js
@@ -2,7 +2,6 @@
 // @match          http://comicalpic.net/img-*.html
 // @match          http://croftimage.com/img-*.html
 // @match          http://hosterbin.com/img-*.html
-// @match          http://image.torrentjav.net/img-*.html
 // @match          http://imagedecode.com/img-*.html
 // @match          http://imgserve.net/img-*.html
 // @match          http://ligasampiona.com/img-*.html
@@ -15,7 +14,7 @@
 
   $register({
     rule: {
-      host: /(imagedecode|zonezeedimage|zeljeimage|ligasampiona|hosterbin|croftimage)\.com|(comicalpic|image\.torrentjav|imgserve)\.net/,
+      host: /(imagedecode|zonezeedimage|zeljeimage|ligasampiona|hosterbin|croftimage)\.com|(comicalpic|imgserve)\.net/,
     },
     run: function () {
       var o = $('#container img[class^=centred]');