+++ src/sites/linkbucks.js
@@ -2,7 +2,6 @@
 // @match          http://*.allanalpass.com/*
 // @match          http://*.amy.gs/*
 // @match          http://*.any.gs/*
-// @match          http://*.baberepublic.com/*
 // @match          http://*.deb.gs/*
 // @match          http://*.drstickyfingers.com/*
 // @match          http://*.dyo.gs/*