+++ src/sites/reklama.js
@@ -1,7 +1,6 @@
 // ==UserScript==
 // @match          http://comicalpic.net/img-*.html
 // @match          http://croftimage.com/img-*.html
-// @match          http://hosterbin.com/img-*.html
 // @match          http://imagedecode.com/img-*.html
 // @match          http://imgserve.net/img-*.html
 // @match          http://ligasampiona.com/img-*.html
@@ -14,7 +13,7 @@
 
   $register({
     rule: {
-      host: /(imagedecode|zonezeedimage|zeljeimage|ligasampiona|hosterbin|croftimage)\.com|(comicalpic|imgserve)\.net/,
+      host: /(imagedecode|zonezeedimage|zeljeimage|ligasampiona|croftimage)\.com|(comicalpic|imgserve)\.net/,
     },
     run: function () {
       var o = $('#container img[class^=centred]');