+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## TBA
+
+* changed
+    * split script to parts
+
 ## 2.25.1
 
 * fixed