+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * gzvd.info
     * image.torrentjav.net
     * youfap.com
+    * baberepublic.com
 * changed
     * split script to parts