+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * 9.bb
     * image69.us
     * jpdown.info
+    * pushba.com
 * changed
     * split script to parts